% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complaints.R
\name{get_complaints}
\alias{get_complaints}
\title{Get complaints by year/make/model}
\usage{
get_complaints(year, make, model)
}
\arguments{
\item{year}{Numeric model year}

\item{make}{Vehicle make}

\item{model}{Vehicle model}
}
\value{
A tibble of complaints
}
\description{
Note: For Tesla Model Y, only 2020 returns complaints; other years return empty.
}
