% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complaints.R
\name{get_complaints_by_odi_number}
\alias{get_complaints_by_odi_number}
\title{Get complaints by ODI number}
\usage{
get_complaints_by_odi_number(odi_number)
}
\arguments{
\item{odi_number}{ODI number}
}
\value{
A tibble of complaints
}
\description{
Get complaints by ODI number
}
