% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratings.R
\name{get_ratings}
\alias{get_ratings}
\title{Get crash test ratings for a specific vehicle}
\usage{
get_ratings(vehicle_id)
}
\arguments{
\item{vehicle_id}{VehicleId from list_variants()}
}
\value{
A tibble of ratings
}
\description{
Get crash test ratings for a specific vehicle
}
