% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratings.R
\name{get_ratings_by_model}
\alias{get_ratings_by_model}
\title{Get ratings for all variants of a year/make/model}
\usage{
get_ratings_by_model(year, make, model)
}
\arguments{
\item{year}{Numeric model year}

\item{make}{Vehicle make}

\item{model}{Vehicle model}
}
\value{
A tibble of ratings for all variants
}
\description{
Get ratings for all variants of a year/make/model
}
