% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recalls.R
\name{get_recalls}
\alias{get_recalls}
\title{Get recalls by year/make/model}
\usage{
get_recalls(year, make, model)
}
\arguments{
\item{year}{Numeric model year}

\item{make}{Vehicle make}

\item{model}{Vehicle model}
}
\value{
A tibble of recalls
}
\description{
Get recalls by year/make/model
}
