% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recalls.R
\name{get_recalls_by_campaign_number}
\alias{get_recalls_by_campaign_number}
\title{Get recalls by campaign number}
\usage{
get_recalls_by_campaign_number(campaign_number)
}
\arguments{
\item{campaign_number}{Campaign number}

\item{year}{Numeric model year}
}
\value{
A tibble of recalls
}
\description{
Get recalls by campaign number
}
