% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_complaints_catalog.R
\name{list_complaint_makes}
\alias{list_complaint_makes}
\title{List all available complaint makes}
\usage{
list_complaint_makes(year)
}
\arguments{
\item{year}{Numeric model year}
}
\value{
A tibble of makes
}
\description{
List all available complaint makes
}
