% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_complaints_catalog.R
\name{list_complaint_models}
\alias{list_complaint_models}
\title{List all available complaint models}
\usage{
list_complaint_models(year, make)
}
\arguments{
\item{year}{Numeric model year}

\item{make}{Vehicle make}
}
\value{
A tibble of models
}
\description{
List all available complaint models
}
