% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_complaints_catalog.R
\name{list_complaint_years}
\alias{list_complaint_years}
\title{List all available complaint years}
\usage{
list_complaint_years()
}
\value{
A tibble of years
}
\description{
List all available complaint years
}
