% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ratings_catalog.R
\name{list_rating_makes}
\alias{list_rating_makes}
\title{List all makes for a given year}
\usage{
list_rating_makes(year)
}
\arguments{
\item{year}{Numeric model year}
}
\value{
A tibble of makes
}
\description{
List all makes for a given year
}
