% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ratings_catalog.R
\name{list_rating_model_variants}
\alias{list_rating_model_variants}
\title{List crash-tested variants for a given year/make/model}
\usage{
list_rating_model_variants(year, make, model)
}
\arguments{
\item{year}{Numeric model year}

\item{make}{Vehicle make}

\item{model}{Vehicle model}
}
\value{
A tibble of vehicle variants with VehicleId
}
\description{
List crash-tested variants for a given year/make/model
}
