% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ratings_catalog.R
\name{list_rating_models}
\alias{list_rating_models}
\title{List all models for a given year and make}
\usage{
list_rating_models(year, make)
}
\arguments{
\item{year}{Numeric model year}

\item{make}{Vehicle make (string)}
}
\value{
A tibble of models
}
\description{
List all models for a given year and make
}
