% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_recalls_catalog.R
\name{list_recall_makes}
\alias{list_recall_makes}
\title{List all available recall makes}
\usage{
list_recall_makes(year)
}
\arguments{
\item{year}{Numeric model year}
}
\value{
A tibble of years
}
\description{
List all available recall makes
}
