% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_recalls_catalog.R
\name{list_recall_models}
\alias{list_recall_models}
\title{List all available recall models}
\usage{
list_recall_models(year, make)
}
\arguments{
\item{year}{Numeric model year}

\item{make}{Vehicle make}
}
\value{
A tibble of years
}
\description{
List all available recall models
}
