% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{make_request}
\alias{make_request}
\title{Perform an HTTP GET request to the NHTSA Safety Ratings API}
\usage{
make_request(path, query = list())
}
\arguments{
\item{endpoint}{API endpoint path (string), e.g., "modelyear/2013/make/Acura/model/RDX"}
}
\value{
Parsed JSON response as a list
}
\description{
Perform an HTTP GET request to the NHTSA Safety Ratings API
}
