% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{parse_response}
\alias{parse_response}
\title{Parse raw JSON response from NHTSA API into a tidy list/tibble}
\usage{
parse_response(response)
}
\arguments{
\item{response}{Raw JSON string}
}
\value{
Parsed response as a list or tibble
}
\description{
Parse raw JSON response from NHTSA API into a tidy list/tibble
}
